+++ src/plugins/externalTool.nim
@@ -8,102 +8,77 @@
 ## This plugin uses information from the config file to set metadata
 ## keys.
 
-import std/algorithm
-import ".."/[config, chalkjson, plugin_api]
+import std/[algorithm, sequtils]
+import ".."/[config, plugin_api, util]
 
-type
-  PIInfo      = ref object
-    name: string
-    obj: ToolInfo
+type PIInfo = ref object
+  name: string
+  obj:  ToolInfo
 
-template broken(cb: CallbackObj, info: PIInfo) =
-  error("missing implementation of " & $(cb) & " for tool: " & info.name)
-  return false
+proc ensureRunCallback[T](cb: CallbackObj, args: seq[Box]): T =
+  let value = runCallback(cb, args)
+  if value.isNone():
+    raise newException(ValueError, "missing implemenetation of " & $(cb))
+  return unpack[T](value.get())
 
-proc runOneTool(info: PIInfo, path: string, dict: var ChalkDict): bool =
-  var
-    args   = @[pack(path)]
-    locbox = runCallback(info.obj.getToolLocation, args)
+var toolCache = initTable[string, ChalkDict]()
+proc runOneTool(info: PIInfo, path: string): ChalkDict =
+  let key = info.name & ":" & path
+  if key in toolCache:
+    return toolCache[key]
 
-  if locBox.isNone(): broken(info.obj.getToolLocation, info)
-  var
-    path   = unpack[string](locBox.get()).strip()
+  trace("Running tool: " & info.name)
+  result = ChalkDict()
+  let args = @[pack(path)]
+  var exe  = ensureRunCallback[string](info.obj.getToolLocation, args)
 
-  if path == "":
-    let installed = runCallback(info.obj.attemptInstall, args)
-    if installed.isNone(): broken(info.obj.attemptInstall, info)
-    if not unpack[bool](installed.get()): return false
-    locbox = runCallback(info.obj.getToolLocation, args)
-    if locBox.isNone(): broken(info.obj.getToolLocation, info)
-    path = unpack[string](locBox.get()).strip()
+  if exe == "":
+    let installed = ensureRunCallback[bool](info.obj.attemptInstall, args)
+    if not installed:
+      trace(info.name & ": could not be installed. skipping")
+      return
+    exe = ensureRunCallback[string](info.obj.getToolLocation, args)
 
-  let argv = runCallback(info.obj.getCommandArgs, args)
-  if argv.isNone(): broken(info.obj.getCommandArgs, info)
+  if exe == "":
+    trace(info.name & ": could not be found. skipping")
+    return
+
   let
-    cmd    = pack(path & " " & unpack[string](argv.get()).strip())
-    outs   = unpack[seq[Box]](c4mSystem(@[cmd]).get())
-
+    argv = ensureRunCallback[string](info.obj.getCommandArgs, args)
+    cmd  = exe & " " & argv.strip()
+  trace(cmd)
   let
-    retOpt = runCallback(info.obj.produceKeys, outs)
-  if retOpt.isNone(): broken(info.obj.produceKeys, info)
-  let
-    d   = unpack[ChalkDict](retOpt.get())
+    outs = unpack[seq[Box]](c4mSystem(@[pack(cmd)]).get())
 
-  if len(d) == 0: return false
+  let d = ensureRunCallback[ChalkDict](info.obj.produceKeys, outs)
+  if len(d) == 0:
+    trace(info.name & ": produced no keys. skipping")
+    return
   if d.contains("error"):
-    error(unpack[string](d["error"]))
-    return false
+    error(info.name & ": " & unpack[string](d["error"]))
+    return
   if d.contains("warn"):
-    warn(unpack[string](d["warn"]))
+    warn(info.name & ": " & unpack[string](d["warn"]))
     d.del("warn")
   if d.contains("info"):
-    warn(unpack[string](d["info"]))
+    info(info.name & ": " & unpack[string](d["info"]))
     d.del("info")
 
-  # The ChalkDict returned is of the form: chalkKey: chalkContents.
-  # For instance, it might come back SBOM : "{arbitrary sbom as a
-  # string}" If the string we get back parses as valid JSON, we will
-  # expand it out.  If not, we use it as-is.
-  #
-  # These k/v pairs will be merged into the dict field, which will take
-  # the form, SBOM : { info.name : chalkContents }
-  #
-  # Where the ChalkContents is either a JSON object or a string
-  # literal.
+  trace(info.name & ": produced keys " & $(d.keys().toSeq()))
+  toolCache[key] = d
+  return d
 
-  for k, v in d:
-    var kindChalkDict: ChalkDict
-    if k in dict:
-      kindChalkDict = unpack[ChalkDict](dict[k])
-    else:
-      kindChalkDict = ChalkDict()
-
-    kindChalkDict[info.name] = v
-    for k, v in kindChalkDict:
-      try:
-        # Attempt to parse as a JSON object.  If not, treat it as a string,
-        # but encode it into a JSON object.
-        let
-          asJson = parseJson(unpack[string](v))
-          asBox  = asJson.nimJsonToBox()
-        kindChalkDict[k] = asBox
-      except:
-        kindChalkDict[k] = pack[string](escapeJson(unpack[string](v)))
-
-    dict[k] = pack(kindChalkDict)
-
-  return info.obj.stopOnSuccess
-
-template toolBase(s: string) {.dirty.} =
+template toolBase(path: string) {.dirty.} =
   result = ChalkDict()
 
   var
-    toolInfo: Table[string, seq[(int, PIInfo)]]
+    toolInfo = initTable[string, seq[(int, PIInfo)]]()
   let
-    runSbom = get[bool](chalkConfig, "run_sbom_tools")
-    runSast = get[bool](chalkConfig, "run_sast_tools")
+    runSbom  = get[bool](chalkConfig, "run_sbom_tools")
+    runSast  = get[bool](chalkConfig, "run_sast_tools")
 
   # tools should only run during insert operations
+  # note this is a subset of chalkable operations
   if getCommandName() notin @["build", "insert"]:
     return result
 
@@ -112,23 +87,33 @@ template toolBase(s: string) {.dirty.} =
     if not runSbom and v.kind == "sbom": continue
     if not runSast and v.kind == "sast": continue
 
-    let priority = v.priority
+    let tool = (v.priority, PIInfo(name: k, obj: v))
     if v.kind notin toolInfo:
-      toolInfo[v.kind] =  @[(priority, PIInfo(name: k, obj: v))]
+      toolInfo[v.kind] = @[tool]
     else:
-      toolInfo[v.kind].add((priority, PIInfo(name: k, obj: v)))
+      toolInfo[v.kind].add(tool)
 
   for k, v in toolInfo:
-    var sortArr = v
-    sortArr.sort()
-    for (ignore, info) in sortArr:
-      trace("Running tool: " & info.name)
-      if info.runOneTool(s, result): break
+    for (ignore, info) in v.sorted():
+      try:
+        let data = info.runOneTool(path)
+        # merge multiple tools into a single structure
+        # for example first tool returns:
+        # { SBOM: { foo: {...} } }
+        # and second tool returns:
+        # { SBOM: { bar: {...} } }
+        # merged structure should be:
+        # { SBOM: { foo: {...}, bar: {...} } }
+        result.merge(data.nestWith(info.name))
+        if len(data) >= 0 and info.obj.stopOnSuccess:
+          break
+      except:
+        error(info.name & ": " & getCurrentExceptionMsg())
 
-proc toolGetChalkTimeHostInfo*(self: Plugin): ChalkDict {.cdecl.} =
+proc toolGetChalkTimeHostInfo(self: Plugin): ChalkDict {.cdecl.} =
   toolBase(resolvePath(getContextDirectories()[0]))
 
-proc toolGetChalkTimeArtifactInfo*(self: Plugin, obj: ChalkObj):
+proc toolGetChalkTimeArtifactInfo(self: Plugin, obj: ChalkObj):
                                  ChalkDict {.cdecl.} =
   if obj.fsRef != "":
     toolbase(resolvePath(obj.fsRef))